aiservice/requirements.txt
gunicorn==20.1.0
albucore==0.0.24
albumentations==2.0.8
alembic==1.16.4
annotated-types==0.7.0
azure-core==1.35.0
azure-storage-blob==12.13.0
blinker==1.9.0
click==8.2.1
colorlog==6.9.0
contourpy==1.3.3
cycler==0.12.1
filetype==1.2.0
Flask==3.1.1
fonttools==4.59.0
fsspec==2025.7.0
greenlet==3.2.4
httpie==3.2.4
idna==3.7
isodate==0.7.2
itsdangerous==2.2.0
kiwisolver==1.4.9
Mako==1.3.10
matplotlib==3.10.5
mpmath==1.3.0
msrest==0.7.1
multidict==6.6.3
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
optuna==4.4.0
pandas==2.3.1
pi_heif==1.1.0
pillow==11.3.0
pillow-avif-plugin==1.5.2
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
PySocks==1.7.1
python-dotenv==1.1.1
pytz==2025.2
requests-oauthlib==2.0.0
roboflow==1.2.3
scipy==1.16.1
simsimd==6.5.0
SQLAlchemy==2.0.42
stringzilla==3.12.6
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typing-inspection==0.4.1
tzdata==2025.2
ultralytics==8.3.176
ultralytics-thop==2.0.15
Werkzeug==3.1.3
